fix(trainer): switch fp16 to bf16 for Ampere to avoid GradScaler bf16 crash
trainer.train() crashed at the first gradient unscale step with
'_amp_foreach_non_finite_check_and_unscale_cuda not implemented for
BFloat16'. Cause: fp16=True engages GradScaler, but transformers/bnb
internally promotes some grads to bfloat16, and GradScaler can't
unscale bf16 tensors. RTX 3060 (Ampere, sm_86) supports bf16 natively
and bf16 doesn't need a scaler at all (same exponent range as fp32),
so use bf16 end-to-end: bnb_4bit_compute_dtype=torch.bfloat16 and
bf16=True in SFTConfig.

+++ backend/train_worker.py
@@ -216,7 +216,7 @@ def run_training(
     bnb_config = BitsAndBytesConfig(
         load_in_4bit=True,
         bnb_4bit_quant_type="nf4",
-        bnb_4bit_compute_dtype=torch.float16,
+        bnb_4bit_compute_dtype=torch.bfloat16,
         bnb_4bit_use_double_quant=True,
     )
 
@@ -271,7 +271,7 @@ def run_training(
         logging_steps=1,
         save_strategy="no",
         report_to=[],
-        fp16=True,
+        bf16=True,
         optim="paged_adamw_8bit",
         dataset_text_field="text",
         max_length=1024,
